clean js utils folder of unused stuff or better located
