+++ app/assets/javascript/js_components/locationFinder/api.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import logger from './logging';
+import logger from '../../lib/logging';
 
 export const getPostcodeFromCoordinates = (latitude, longitude) => axios.get('https://api.postcodes.io/postcodes', {
   params: { latitude, longitude },
@@ -8,17 +8,8 @@ export const getPostcodeFromCoordinates = (latitude, longitude) => axios.get('ht
     logger.log(`${error} Postcodes API`);
   });
 
-export const getLocationSuggestions = ({ query, populateResults }) => axios.get(`/api/v1/location_suggestion/${query}?format=json`)
-  .then((response) => response.data)
-  .then((data) => data.suggestions)
-  .then(populateResults)
-  .catch((error) => {
-    logger.log(`${error} Search query: ${query}`);
-  });
-
 const api = {
   getPostcodeFromCoordinates,
-  getLocationSuggestions,
 };
 
 export default api;
