--- app/assets/javascript/lib/__mocks__/utils.js
@@ -1,1 +1,0 @@
-export const updateUrlQueryParams = jest.fn();
